Added cast around GLEW_ARB_texture_non_power_of_two, as it might be a char and screw up output
SVN-Revision: 6231

+++ src/video/gl/gl_renderer.cpp
@@ -180,7 +180,7 @@ GLRenderer::GLRenderer() :
     throw std::runtime_error(out.str());
   }
   log_info << "Using GLEW " << glewGetString(GLEW_VERSION) << std::endl;
-  log_info << "GLEW_ARB_texture_non_power_of_two: " << GLEW_ARB_texture_non_power_of_two << std::endl;
+  log_info << "GLEW_ARB_texture_non_power_of_two: " << static_cast<int>(GLEW_ARB_texture_non_power_of_two) << std::endl;
 #endif
 }
